+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Added
+- custom::sql block for executing SQL statements
+
 ### Fixed
 - Some codesmells according to SonarCloud